Data.ByteString.Lazy.Base32

Description

This module contains ByteString-valued combinators for implementing the RFC 4648 specification of the Base32 encoding format. This includes strictly padded/unpadded decoding variants, as well as internal and external validation for canonicity.

Encoding

encodeBase32 :: ByteString -> Text

Encode a ByteString value as a Base32 Text value with padding.

See: RFC-4648 section 6

Examples:

>>> encodeBase32 "Sun"
"KN2W4==="

encodeBase32' :: ByteString -> ByteString

Encode a ByteString as a Base32 ByteString value with padding.

See: RFC-4648 section 6

Examples:

>>> encodeBase32 "Sun"
"KN2W4==="

encodeBase32Unpadded :: ByteString -> Text

Encode a ByteString value as Base32 Text without padding. Note that for Base32, padding is optional. If you call this function, you will simply be encoding as Base32 and stripping padding chars from the output.

See: RFC-4648 section 6

Examples:

>>> encodeBase32Unpadded "Sun"
"KN2W4"

encodeBase32Unpadded' :: ByteString -> ByteString

Encode a ByteString value as Base32 without padding. Note that for Base32, padding is optional. If you call this function, you will simply be encoding as Base32 and stripping padding chars from the output.

See: RFC-4648 section 6

Examples:

>>> encodeBase32Unpadded' "Sun"
"KN2W4"

Decoding

decodeBase32 :: ByteString -> Either Text ByteString

Decode an arbitrarily padded Base32 encoded ByteString value. If its length is not a multiple of 4, then padding chars will be added to fill out the input to a multiple of 4 for safe decoding as Base32-encoded values are optionally padded.

See: RFC-4648 section 6

Examples:

>>> decodeBase32 "KN2W4==="
Right "Sun"

>>> decodeBase32 "KN2W4"
Right "Sun"

>>> decodeBase32 "KN2W==="
Left "Base32-encoded bytestring has invalid padding"

decodeBase32Unpadded :: ByteString -> Either Text ByteString

Decode an unpadded Base32-encoded ByteString value. Input strings are required to be unpadded, and will undergo validation prior to decoding to confirm.

In general, unless unpadded Base32 is explicitly required, it is safer to call decodeBase32.

See: RFC-4648 section 6

Examples:

>>> decodeBase32Unpadded "KN2W4"
Right "Sun"

>>> decodeBase32Unpadded "KN2W4==="
Left "Base32-encoded bytestring has invalid padding"

decodeBase32Padded :: ByteString -> Either Text ByteString

Decode a padded Base32-encoded ByteString value. Input strings are required to be correctly padded, and will be validated prior to decoding to confirm.

In general, unless padded Base32 is explicitly required, it is safer to call decodeBase32.

See: RFC-4648 section 6

Examples:

>>> decodeBase32Padded "KN2W4==="
Right "Sun"

>>> decodeBase32Padded "KN2W4"
Left "Base32-encoded bytestring requires padding"

Validation

isBase32 :: ByteString -> Bool

Tell whether a ByteString is Base32-encoded.

Examples:

>>> isBase32 "KN2W4"
True

>>> isBase32 "KN2W4==="
True

>>> isBase32 "KN2W4=="
False

isValidBase32 :: ByteString -> Bool

Tell whether a ByteString is a valid Base32 format.

This will not tell you whether or not this is a correct Base32 representation, only that it conforms to the correct shape. To check whether it is a true Base32 encoded ByteString value, use isBase32.

Examples:

>>> isValidBase32 "KN2W4"
True

>>> isValidBase32 "KN2W4="
False

>>> isValidBase32 "KN2W4%"
False