Data.ByteString.Short.Base32.Hex

Description

This module contains ShortByteString-valued combinators for implementing the RFC 4648 specification of the Base32hex encoding format. This includes strictly padded/unpadded and decoding variants, as well as internal and external validation for canonicity.

Encoding

encodeBase32 :: ShortByteString -> ShortText

Encode a ShortByteString value as a Base32hex Text value with padding.

See: RFC-4648 section 7

Examples:

>>> encodeBase32 "Sun"
"ADQMS==="

encodeBase32' :: ShortByteString -> ShortByteString

Encode a ShortByteString as a Base32hex ShortByteString value with padding.

See: RFC-4648 section 7

Examples:

>>> encodeBase32' "Sun"
"ADQMS==="

encodeBase32Unpadded :: ShortByteString -> ShortText

Encode a ShortByteString value as Base32hex Text without padding. Note that for Base32hex, padding is optional. If you call this function, you will simply be encoding as Base32hex and stripping padding chars from the output.

See: RFC-4648 section 7

Examples:

>>> encodeBase32Unpadded' "Sun"
"ADQMS"

encodeBase32Unpadded' :: ShortByteString -> ShortByteString

Encode a ShortByteString value as Base32hex without padding. Note that for Base32hex, padding is optional. If you call this function, you will simply be encoding as Base32hex and stripping padding chars from the output.

See: RFC-4648 section 7

Examples:

>>> encodeBase32Unpadded' "Sun"
"ADQMS"

Decoding

decodeBase32 :: ShortByteString -> Either Text ShortByteString

Decode an arbitrarily padded Base32hex encoded ShortByteString value. If its length is not a multiple of 4, then padding chars will be added to fill out the input to a multiple of 4 for safe decoding as Base32hex-encoded values are optionally padded.

See: RFC-4648 section 7

Examples:

>>> decodeBase32 "ADQMS==="
Right "Sun"

>>> decodeBase32 "ADQMS"
Right "Sun"

>>> decodeBase32 "ADQM==="
Left "Base32-encoded bytestring has invalid padding"

decodeBase32Unpadded :: ShortByteString -> Either Text ShortByteString

Decode an unpadded Base32hex-encoded ShortByteString value. Input strings are required to be unpadded, and will undergo validation prior to decoding to confirm.

In general, unless unpadded Base32hex is explicitly required, it is safer to call decodeBase32.

See: RFC-4648 section 7

Examples:

>>> decodeBase32Unpadded "ADQMS"
Right "Sun"

>>> decodeBase32Unpadded "ADQMS==="
Left "Base32-encoded bytestring has invalid padding"

decodeBase32Padded :: ShortByteString -> Either Text ShortByteString

Decode a padded Base32hex-encoded ShortByteString value. Input strings are required to be correctly padded, and will be validated prior to decoding to confirm.

In general, unless padded Base32hex is explicitly required, it is safer to call decodeBase32.

See: RFC-4648 section 7

Examples:

>>> decodeBase32Padded "ADQMS==="
Right "Sun"

>>> decodeBase32Padded "ADQMS"
Left "Base32-encoded bytestring requires padding"

Validation

isBase32Hex :: ShortByteString -> Bool

Tell whether a ShortByteString is Base32hex-encoded.

Examples:

>>> isBase32Hex "ADQMS"
True

>>> isBase32Hex "ADQMS==="
True

>>> isBase32Hex "ADQMS=="
False

isValidBase32Hex :: ShortByteString -> Bool

Tell whether a ShortByteString is a valid Base32hex format.

This will not tell you whether or not this is a correct Base32hex representation, only that it conforms to the correct shape. To check whether it is a true Base32 encoded ShortByteString value, use isBase32Hex.

Examples:

>>> isValidBase32Hex "ADQMS"
True

>>> isValidBase32Hex "ADQMS="
False

>>> isValidBase32Hex "ADQMS%"
False