Data.Text.Encoding.Base32

Description

This module contains the combinators implementing the RFC 4648 specification for the Base32 encoding including unpadded and lenient variants

Documentation

encodeBase32 :: Text -> Text

Encode a Text value in Base32 with padding.

See: RFC-4648 section 6

decodeBase32 :: Text -> Either Text Text

Decode a padded Base32-encoded Text value

See: RFC-4648 section 6

encodeBase32Unpadded :: Text -> Text

Encode a Text value in Base32 without padding.

See: RFC-4648 section 6, RFC-4648 section 3.2

decodeBase32Unpadded :: Text -> Either Text Text

Decode an arbitrarily padded Base32-encoded Text

See: RFC-4648 section 6, RFC-4648 section 3.2

isBase32 :: Text -> Bool

Tell whether a Text value is Base32-encoded.

isValidBase32 :: Text -> Bool

Tell whether a Text value is a valid Base32 format.

This will not tell you whether or not this is a correct Base32 representation, only that it conforms to the correct shape. To check whether it is a true Base32 encoded Text value, use isBase32.