Data.Text.Encoding.Base32.Error

Description

This module contains the error types raised (not as exceptions!) in the decoding process.

Documentation

data Base32Error e

This data type represents the type of decoding errors of various kinds as they pertain to decoding Text values. Namely, to distinguish between decoding errors from opaque unicode exceptions caught in the unicode decoding process.

Constructors

DecodeError Text
The error associated with decoding failure as a result of the Base32 decoding process

ConversionError e
The error associated with the decoding failure as a result of the conversion process