Data.Text.Short.Encoding.Base32

Description

This module contains ShortText-valued combinators implementing the RFC 4648 specification for the Base32 encoding format. This includes strictly padded/unpadded decoding variants, and external + internal validations for canonicity.

Encoding

encodeBase32 :: ShortText -> ShortText

Encode a ShortText value in Base32 with padding.

See: RFC-4328 section 6

Examples:

>>> encodeBase32 "Sun"
"KN2W4==="

encodeBase32Unpadded :: ShortText -> ShortText

Encode a ShortText value in Base32 without padding. Note that for Base32, padding is optional. If you call this function, you will simply be encoding as Base32 and stripping padding chars from the output.

See: RFC-4328 section 6

Examples:

>>> encodeBase32Unpadded "Sun"
"KN2W4"

Decoding

decodeBase32 :: ShortText -> Either Text ShortText

Decode an arbitrarily padded Base32-encoded ShortText value. If its length is not a multiple of 4, then padding chars will be added to fill out the input to a multiple of 4 for safe decoding as base32 encodings are optionally padded.

Note: This function makes sure that decoding is total by deferring to decodeUtf8. This will always round trip for any valid Base32-encoded text value, but it may not round trip for bad inputs. The onus is on the caller to make sure inputs are valid. If unsure, defer to decodeBase32With and pass in a custom decode function.

See: RFC-4328 section 6

Examples:

>>> decodeBase32 "KN2W4==="
Right "Sun"

>>> decodeBase32 "KN2W4"
Right "Sun"

>>> decodeBase32 "KN2W==="
Left "Base32-encoded bytestring has invalid padding"

decodeBase32With

Arguments

:: (ShortByteString -> Either err ShortText)
convert a bytestring to text (e.g. decodeUtf8')

-> ShortByteString
Input text to decode

-> Either (Base32Error err) ShortText

Attempt to decode a ShortByteString value as Base32, converting from ByteString to ShortText according to some encoding function. In practice, This is something like decodeUtf8', which may produce an error.

See: RFC-4328 section 6

Examples:

decodeBase32With '(fmap fromText . T.decodeUtf8' . toText)'
  :: ShortByteString -> Either (Base32Error UnicodeException) ShortText

decodeBase32Unpadded :: ShortText -> Either Text ShortText

Decode an unpadded Base32 encoded ShortText value.

Note: This function makes sure that decoding is total by deferring to decodeUtf8. This will always round trip for any valid Base32-encoded text value, but it may not round trip for bad inputs. The onus is on the caller to make sure inputs are valid. If unsure, defer to decodeBase32UnpaddedWith and pass in a custom decode function.

See: RFC-4328 section 6

Examples:

>>> decodeBase32Unpadded "KN2W4"
Right "Sun"

>>> decodeBase32Unpadded "KN2W4==="
Left "Base32-encoded bytestring has invalid padding"

decodeBase32UnpaddedWith

Arguments

:: (ShortByteString -> Either err ShortText)
convert a bytestring to text (e.g. decodeUtf8')

-> ShortByteString
Input text to decode

-> Either (Base32Error err) ShortText

Attempt to decode an unpadded ShortByteString value as Base32, converting from ShortByteString to ShortText according to some encoding function. In practice, This is something like decodeUtf8', which may produce an error.

See: RFC-4328 section 6

Examples:

decodeBase32UnpaddedWith '(fmap fromText . T.decodeUtf8' . toText)'
  :: ShortByteString -> Either (Base32Error UnicodeException) ShortText

decodeBase32Padded :: ShortText -> Either Text ShortText

Decode an padded Base32 encoded ShortText value

Note: This function makes sure that decoding is total by deferring to decodeUtf8. This will always round trip for any valid Base32-encoded text value, but it may not round trip for bad inputs. The onus is on the caller to make sure inputs are valid. If unsure, defer to decodeBase32PaddedWith and pass in a custom decode function.

See: RFC-4328 section 6

Examples:

>>> decodeBase32Padded "KN2W4==="
Right "Sun"

>>> decodeBase32Padded "KN2W4"
Left "Base32-encoded bytestring requires padding"

decodeBase32PaddedWith

Arguments

:: (ShortByteString -> Either err ShortText)
convert a bytestring to text (e.g. decodeUtf8')

-> ShortByteString
Input text to decode

-> Either (Base32Error err) ShortText

Attempt to decode a padded ShortByteString value as Base32, converting from ByteString to ShortText according to some encoding function. In practice, This is something like decodeUtf8', which may produce an error.

See: RFC-4328 section 6

Examples:

decodeBase32With '(fmap fromText . T.decodeUtf8' . toText)'
  :: ShortByteString -> Either (Base32Error UnicodeException) ShortText

Validation

isBase32 :: ShortText -> Bool

Tell whether a ShortText value is Base32-encoded.

Examples:

>>> isBase32 "KN2W4"
True

>>> isBase32 "KN2W4==="
True

>>> isBase32 "KN2W4=="
False

isValidBase32 :: ShortText -> Bool

Tell whether a ShortText value is a valid Base32 format.

This will not tell you whether or not this is a correct Base32 representation, only that it conforms to the correct shape. To check whether it is a true Base32 encoded ShortText value, use isBase32.

Examples:

>>> isValidBase32 "KN2W4"
True

>>> isValidBase32 "KN2W4="
False

>>> isValidBase32 "KN2W4%"
False